Descriptor Management
Created The Committed Resource heap(CRH).
Users can use it for the initial creation of assets, this is where the resources and corresponding descriptors will live.

Allocate functions will return Handles of the corresponding type.  these handles can be translated by the (CRH).

Further functionallity will be added in the future, the CRH. It will be used as a source for the staging descriptor heaps.